chores: organizing and moving around state/layout parse methods
